Partial overhaul of main flashfocus classes
Plan is to separate the flashing and signal handling functions of the
server. Still need to add a lot of tests and documentation but the main
structure is there. Also planning to add the 'default_opacity'
paramater. This will simplify the threading code significantly because
windows won't have to be locked. The only downside is that users who
have a default opacity != 1 will have to set another parameter.